ptrdiff_t

+++ function.h
@@ -35,9 +35,9 @@ class	CStatement
 {
 public:
 	int				type;			// ステートメントの種別
-	int				jumpto;			// 飛び先行番号 break/continue/return/if/elseif/else/for/foreachで使用します
+	ptrdiff_t		jumpto;			// 飛び先行番号 break/continue/return/if/elseif/else/for/foreachで使用します
 									// 該当単位終端の"}"の位置が格納されています
-	int	linecount;					// 辞書ファイル中の行番号
+	ptrdiff_t linecount;			// 辞書ファイル中の行番号
 
 	mutable std_shared_ptr < CDuplEvInfo >	dupl_block;		// pool:{ //...
 
@@ -46,7 +46,7 @@ private:
 	mutable std_shared_ptr<std::vector<CSerial> >	m_serial;			// 数式の演算順序
 
 public:
-	CStatement(int t, int l, std_shared_ptr<CDuplEvInfo> dupl = std_shared_ptr<CDuplEvInfo>() )
+	CStatement(int t, ptrdiff_t l, std_shared_ptr<CDuplEvInfo> dupl = std_shared_ptr<CDuplEvInfo>() )
 	{
 		type = t;
 		linecount = l;
